backend/requirements.txt
asgiref==3.8.1
django==5.1.3
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
supabase>=2.0.0
djangorestframework
whitenoise==4.1.2
markdown
django-filter
PyYAML
django-guardian
django-cors-headers
pgvector
azure-cognitiveservices-speech
openai
pydub
audioop-lts